refactor: unify sitemap visibility with isPostVisible
Sitemap exclusion derived from hiddenPostUrls treated draft:true as
hidden unconditionally and built /slug/ (trailing slash), drifting from
the runtime isPostVisible policy (draft hidden only in PROD) and risking
silent slash-mismatch re-inclusion.

- isPostVisible takes optional { isProd } so the build-time scanner reuses
  the exact same policy (import.meta.env-safe outside Vite)
- getHiddenPostPaths (was getHiddenPostUrls) applies isPostVisible and
  returns canonical no-trailing-slash paths via shared postPath helper
- postPath: single path-normalization rule, used by both the scanner and
  urlOutsideOfPagesDirPlugin
- sitemap filter strips trailing slash on both sides before lookup
